Better Component Design
