+++ src/OnboardingSPA/components/CardWithOptionsTemplate/OptionItem/stylesheet.scss
@@ -1,0 +1,47 @@
+$background-color: var(--nfd-onboarding-card-background);
+
+.nfd-sg-card__data__option {
+
+	&__wrapper {
+		display: flex;
+		cursor: pointer;
+		margin: 16px 4px;
+		align-items: center;
+		border-radius: 8px;
+		padding: 16px 12px 16px 12px;
+		justify-content: space-between;
+		transition: background-color 400ms ease-in-out;
+
+		&--selected {
+			background-color: var(--nfd-onboarding-hover);
+		}
+	}
+
+	&:not(:last-child) {
+		border-bottom: 0.5px solid rgba(var(--nfd-onboarding-primary-rgb), 0.3);
+	}
+
+	&__left_top {
+		color: var(--nfd-onboarding-primary);
+		font-size: 15px;
+		font-weight: 500;
+		margin-bottom: 7px;
+		line-height: 24px;
+	}
+
+	&__left_bottom {
+		color: rgba(var(--nfd-onboarding-primary-rgb), 0.7);
+
+		font-size: 14px;
+		font-weight: 300;
+	}
+
+	&__right {
+		fill: var(--nfd-onboarding-primary);
+		transition: all 200ms ease-in-out;
+
+		&:hover {
+			transform: scale(1.1);
+		}
+	}
+}
