+++ src/OnboardingSPA/components/CardWithOptionsTemplate/index.js
@@ -1,0 +1,2 @@
+export { default as OptionItem } from './OptionItem';
+export { default as CardWithOptions } from './CardWithOptions';
