+++ src/OnboardingSPA/steps/SiteGen/Experience/index.js
@@ -5,7 +5,7 @@ import getContents from './contents';
 import { HEADER_SITEGEN } from '../../../../constants';
 import { store as nfdOnboardingStore } from '../../../store';
 import CommonLayout from '../../../components/Layouts/Common';
-import CardWithOptions from '../../../components/CardWithOptions';
+import { CardWithOptions } from '../../../components/CardWithOptionsTemplate';
 import SiteGenLoader from '../../../components/Loaders/SiteGenLoader';
 import SitegenAiStateHandler from '../../../components/StateHandlers/SitegenAi';
 
@@ -62,7 +62,6 @@ const SiteGenExperience = () => {
 					<CardWithOptions
 						title={ content.heading }
 						options={ content.options }
-						skip={ content.skip }
 						selection={ selection }
 						callback={ checkAndNavigate }
 					/>
