+++ src/OnboardingSPA/styles/app.scss
@@ -56,7 +56,8 @@
 @import "../components/TextInput/TextInputSiteGen/stylesheet";
 @import "../components/ToggleDarkMode/stylesheet";
 @import "../components/Button/NextButtonSiteGen/stylesheet";
-@import "../components/CardWithOptions/stylesheet";
+@import "../components/CardWithOptionsTemplate/OptionItem/stylesheet";
+@import "../components/CardWithOptionsTemplate/CardWithOptions/stylesheet";
 @import "../components/Loaders/SiteGenLoader/stylesheet";
 @import "../components/Footer/stylesheet";
 @import "../components/LivePreview/SiteGenCard/stylesheet";
